# generator/docs/swagger.yaml
basePath: /api
definitions:
  handlers.JobResponse:
    properties:
      config:
        $ref: '#/definitions/job.JobConfiguration'
      id:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  handlers.PreviewRequest:
    properties:
      rule:
        type: string
      type:
        type: string
    type: object
  handlers.PreviewResponse:
    properties:
      data:
        type: string
    type: object
  job.Job:
    properties:
      config:
        $ref: '#/definitions/job.JobConfiguration'
      id:
        type: string
      name:
        type: string
      stats:
        $ref: '#/definitions/job.Stats'
      status:
        type: string
    type: object
  job.JobConfiguration:
    properties:
      name:
        type: string
      observer:
        items:
          $ref: '#/definitions/observer.Configuration'
        type: array
      sinks:
        items:
          $ref: '#/definitions/sink.Configuration'
        type: array
      source:
        $ref: '#/definitions/source.Configuration'
      timeout:
        type: integer
    type: object
  job.Stats:
    properties:
      failed_write:
        type: integer
      success_write:
        type: integer
    type: object
  observer.Configuration:
    properties:
      properties:
        additionalProperties: true
        type: object
      type:
        type: string
    type: object
  sink.Configuration:
    properties:
      properties:
        additionalProperties: true
        type: object
      type:
        type: string
    type: object
  source.Configuration:
    properties:
      batch_number:
        type: integer
      batch_size:
        type: integer
      concurency:
        type: integer
      fields:
        items:
          $ref: '#/definitions/source.Field'
        type: array
      interval:
        type: integer
      interval_delta:
        type: integer
      random_event:
        type: boolean
    type: object
  source.Field:
    properties:
      limit:
        items: {}
        type: array
      name:
        type: string
      range:
        items: {}
        type: array
      rule:
        type: string
      timestamp_delay_max:
        type: integer
      timestamp_delay_min:
        type: integer
      timestamp_format:
        type: string
      timestamp_locale:
        type: string
      type:
        type: string
    type: object
info:
  contact:
    email: gang@timeplus.io
  description: This is timeplus data generator api server.
  title: Chameleon Generator
  version: "1.0"
paths:
  /jobs:
    get:
      consumes:
      - application/json
      description: list all jobs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/job.Job'
            type: array
      summary: list all jobs.
      tags:
      - job
    post:
      consumes:
      - application/json
      description: create a job.
      parameters:
      - description: job configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/job.JobConfiguration'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.JobResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a job.
      tags:
      - job
  /jobs/{id}:
    delete:
      consumes:
      - application/json
      description: delete job by id.
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      summary: delete job by id.
      tags:
      - job
    get:
      consumes:
      - application/json
      description: get job by id.
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.Job'
        "404":
          description: Not Found
      summary: get job by id.
      tags:
      - job
  /jobs/{id}/start:
    post:
      consumes:
      - application/json
      description: start to run a job.
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      summary: start to run a job.
      tags:
      - job
  /jobs/{id}/stop:
    post:
      consumes:
      - application/json
      description: stop a running job.
      parameters:
      - description: job id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      summary: stop a running job.
      tags:
      - job
  /previews:
    post:
      consumes:
      - application/json
      description: Preview a generated data.
      parameters:
      - description: preview request
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/handlers.PreviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PreviewResponse'
        "400":
          description: Bad Request
      summary: Preview a generated data.
      tags:
      - preview
swagger: "2.0"